API Changes and formatting
