+++ backends/api/prisma.config.ts
@@ -1,8 +1,8 @@
-import path from 'node:path'
+import path from "node:path";
 import { defineConfig, env } from "prisma/config";
 
 export default defineConfig({
-  schema: path.join('prisma'),
+  schema: path.join("prisma"),
   migrations: {
     path: "prisma/migrations",
   },
