+++ backends/api/src/app.ts
@@ -1,10 +1,10 @@
 import express from "express";
 import authRouter from "./auth/routes.ts";
-import vesselRouter from "./vessels/routes.ts";
+import logbookRouter from "./logbooks/routes.ts";
 
 const app = express();
 
 app.use("/auth", authRouter);
-app.use("/vessels", vesselRouter);
+app.use("/logbooks", logbookRouter);
 
 export default app;
