+++ backends/api/src/logbooks/service.ts
@@ -1,0 +1,1 @@
+import { prisma } from "../configs/db.ts";
